Week-2 (changes done)
