+++ app/destination/page.js
@@ -5,6 +5,7 @@ import { useState } from "react";
 import styles from "@/components/destination/destination.module.css";
 import { AddWishlistItem } from "@/components/destination/AddWishlistItem";
 import { planetWishlistItem } from "@/components/destination/planetWishListItem";
+import planetsNames from "@/app/destination/planetsData";
 
 // TASK - React 1 week 2
 // Move this to its own file
@@ -51,32 +52,7 @@ const PlanetCard = ({
     </div>
   );
 };
-const planetsNames = [
-  {
-    id: "europa",
-    name: "Europa",
-    description:
-      "Europa, one of Jupiter’s moons, is an icy world with a hidden ocean beneath its surface. This mysterious moon is a prime candidate for the search for extraterrestrial life, making it a thrilling destination for space explorers.",
-  },
-  {
-    id: "mars",
-    name: "Mars",
-    description:
-      "Mars, the Red Planet, is a barren yet fascinating world with vast deserts, towering volcanoes, and the deepest canyon in the solar system. As humanity’s next frontier, Mars invites us to dream of colonization and the possibilities of life beyond Earth.",
-  },
-  {
-    id: "moon",
-    name: "Moon",
-    description:
-      "Our closest celestial neighbor, the Moon, is a silent witness to Earth's history. With its stunning craters and desolate landscapes, the Moon offers a unique glimpse into space exploration's past and future, making it a perfect destination for lunar adventurers.",
-  },
-  {
-    id: "titan",
-    name: "Titan",
-    description:
-      "Titan, Saturn's largest moon, is a world of dense atmosphere and liquid methane lakes. This enigmatic moon is shrouded in a thick orange haze, concealing a landscape that is both alien and strangely familiar, beckoning explorers to uncover its secrets.",
-  },
-];
+
 export const Destinations = () => {
   const [selectedPlanets, onAddPlanet] = useState([]);
   const numberOfPlanets = selectedPlanets.length;
@@ -92,10 +68,10 @@ export const Destinations = () => {
     } else {
       updatedSelection = [...selectedPlanets, name];
     }
-    onAddPlanet(updatedSelection); // update state always
-    console.log(
+    onAddPlanet(updatedSelection);
+    /*console.log(
       `You selected the following planet: ${name}, with index: ${index}`
-    );
+    );*/
   };
 
   // TASK - React 1 week 2
